backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
langchain>=0.3.7
langchain-core>=0.3.23
langchain-openai>=0.2.9
langchain-anthropic>=0.2.4
langchain-google-genai>=2.0.5
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
langchain-qdrant>=0.1.4
langgraph>=0.2.53
langgraph-checkpoint>=2.0.6
langgraph-prebuilt>=0.1.3
langgraph-sdk>=0.1.42
langsmith>=0.1.140
qdrant-client==1.11.0
youtube-transcript-api==1.2.4
cachetools==5.3.3
google-api-python-client==2.122.0
